githup-api/requirements.txt
langchain>=0.1.0
langgraph>=0.0.30
PyGithub>=1.59.0
openai>=1.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
requests>=2.30.0
fastapi>=0.100.0
uvicorn>=0.20.0
